# docs/swagger.yaml
definitions:
  business.AddBusinessRequest:
    properties:
      description:
        type: string
      email:
        type: string
      gstIndexNumber:
        type: string
      id:
        type: string
      industry:
        $ref: '#/definitions/business.Industry'
      location:
        $ref: '#/definitions/business.Location'
      name:
        type: string
      phone:
        type: string
      pictureUrls:
        items:
          type: string
        type: array
    type: object
...
  business.Industry:
    enum:
    - RetailAndEcommerce
    - FoodAndBeverages
    - HealthAndWellness
    - EducationAndTraining
    - ProfessionalServices
    - Manufacturing
    - HospitalityAndTourism
    - ArtsAndEntertainment
    - TechnologyAndSoftware
    - ConstructionAndRealEstate
    - TransportationAndLogistics
    - AgricultureAndFarming
    - FinanceAndInsurance
    - EnergyAndUtilities
    - NonProfitAndSocialEnterprise
    - MediaAndPublishing
    - Automotive
    - FashionAndLifestyle
    - SportsAndRecreation
    - Other
    type: string
    x-enum-varnames:
    - IndustryRetailAndEcommerce
    - IndustryFoodAndBeverages
    - IndustryHealthAndWellness
    - IndustryEducationAndTraining
    - IndustryProfessionalServices
    - IndustryManufacturing
    - IndustryHospitalityAndTourism
    - IndustryArtsAndEntertainment
    - IndustryTechnologyAndSoftware
    - IndustryConstructionAndRealEstate
    - IndustryTransportationAndLogistics
    - IndustryAgricultureAndFarming
    - IndustryFinanceAndInsurance
    - IndustryEnergyAndUtilities
    - IndustryNonProfitAndSocialEnterprise
    - IndustryMediaAndPublishing
    - IndustryAutomotive
    - IndustryFashionAndLifestyle
    - IndustrySportsAndRecreation
    - IndustryOther
  business.ListJobsResponse:
    properties:
      applicationIds:
        items:
          type: string
        type: array
      description:
        type: string
      eligibility:
        $ref: '#/definitions/job.Eligibility'
      id:
        type: string
      location:
        $ref: '#/definitions/job.Location'
      name:
        type: string
      salaryRange:
        $ref: '#/definitions/job.SalaryRange'
      type:
        $ref: '#/definitions/job.JobType'
      vacancies:
        type: integer
      workDays:
        $ref: '#/definitions/job.WorkDays'
      workHours:
        $ref: '#/definitions/job.WorkHours'
    type: object
  business.Location:
    properties:
      address:
        type: string
      areaCode:
        description: 'Postal code, for example: ''560102'''
        type: string
      city:
        description: 'STD code, for example: ''std:080'''
        type: string
      coordinates:
        $ref: '#/definitions/business.Coordinates'
      state:
        description: 'State code, for example: ''IN-KA'''
        type: string
      street:
        type: string
    type: object
  github_com_ONEST-Network_Job-Manager-Adapter_pkg_database_mongodb_job-application.ApplicantDetails:
    properties:
      age:
        type: integer
      documents:
        $ref: '#/definitions/github_com_ONEST-Network_Job-Manager-Adapter_pkg_database_mongodb_job-application.Documents'
      email:
        type: string
      experience:
        $ref: '#/definitions/github_com_ONEST-Network_Job-Manager-Adapter_pkg_database_mongodb_job-application.Experience'
      gender:
        type: string
      name:
        type: string
      phone:
        type: string
    type: object
...
  github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_cancel_request.Context:
    properties:
      action:
        type: string
      bap_id:
        type: string
      bap_uri:
        type: string
      bpp_id:
        type: string
      bpp_uri:
        type: string
      domain:
        type: string
      location:
        $ref: '#/definitions/github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_cancel_request.Location'
      message_id:
        type: string
      timestamp:
        type: string
      transaction_id:
        type: string
      ttl:
        type: string
      version:
        type: string
    type: object
...
  github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_cancel_request.Message:
    properties:
      cancellation_reason_id:
        type: string
      order_id:
        type: string
    type: object
...
  github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_cancel_response.Context:
    properties:
      action:
        type: string
      bap_id:
        type: string
      bap_uri:
        type: string
      bpp_id:
        type: string
      bpp_uri:
        type: string
      domain:
        type: string
      location:
        $ref: '#/definitions/github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_cancel_response.Location'
      message_id:
        type: string
      timestamp:
        type: string
      transaction_id:
        type: string
      ttl:
        type: string
      version:
        type: string
    type: object
...
  github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_cancel_response.Message:
    properties:
      order:
        $ref: '#/definitions/github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_cancel_response.Order'
    type: object
...
  github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_confirm_request.Context:
    properties:
      action:
        type: string
      bap_id:
        type: string
      bap_uri:
        type: string
      bpp_id:
        type: string
      bpp_uri:
        type: string
      domain:
        type: string
      location:
        $ref: '#/definitions/github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_confirm_request.Location'
      message_id:
        type: string
      timestamp:
        type: string
      transaction_id:
        type: string
      ttl:
        type: string
      version:
        type: string
    type: object
...
  github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_confirm_request.Message:
    properties:
      order:
        $ref: '#/definitions/github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_confirm_request.Order'
    type: object
...
  github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_confirm_response.Context:
    properties:
      action:
        type: string
      bap_id:
        type: string
      bap_uri:
        type: string
      bpp_id:
        type: string
      bpp_uri:
        type: string
      domain:
        type: string
      location:
        $ref: '#/definitions/github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_confirm_response.Location'
      message_id:
        type: string
      timestamp:
        type: string
      transaction_id:
        type: string
      ttl:
        type: string
      version:
        type: string
    type: object
...
  github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_confirm_response.Message:
    properties:
      order:
        $ref: '#/definitions/github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_confirm_response.Order'
    type: object
...
  github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_init_request.Context:
    properties:
      action:
        type: string
      bap_id:
        type: string
      bap_uri:
        type: string
      bpp_id:
        type: string
      bpp_uri:
        type: string
      domain:
        type: string
      location:
        $ref: '#/definitions/github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_init_request.Location'
      message_id:
        type: string
      timestamp:
        type: string
      transaction_id:
        type: string
      ttl:
        type: string
      version:
        type: string
    type: object
...
  github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_init_request.Message:
    properties:
      order:
        $ref: '#/definitions/github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_init_request.Order'
    type: object
...
  github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_init_response.Context:
    properties:
      action:
        type: string
      bap_id:
        type: string
      bap_uri:
        type: string
      bpp_id:
        type: string
      bpp_uri:
        type: string
      domain:
        type: string
      location:
        $ref: '#/definitions/github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_init_response.Location'
      message_id:
        type: string
      timestamp:
        type: string
      transaction_id:
        type: string
      ttl:
        type: string
      version:
        type: string
    type: object
...
  github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_init_response.Message:
    properties:
      order:
        $ref: '#/definitions/github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_init_response.Order'
    type: object
...
  github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_search_request.Context:
    properties:
      action:
        type: string
      bap_id:
        type: string
      bap_uri:
        type: string
      domain:
        type: string
      location:
        $ref: '#/definitions/github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_search_request.Location'
      message_id:
        type: string
      timestamp:
        type: string
      transaction_id:
        type: string
      ttl:
        type: string
      version:
        type: string
    type: object
...
  github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_search_request.Message:
    properties:
      intent:
        $ref: '#/definitions/request.Intent'
    type: object
...
  github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_search_response.Context:
    properties:
      action:
        type: string
      bap_id:
        type: string
      bap_uri:
        type: string
      bpp_id:
        type: string
      bpp_uri:
        type: string
      domain:
        type: string
      location:
        $ref: '#/definitions/github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_search_response.Location'
      message_id:
        type: string
      timestamp:
        type: string
      transaction_id:
        type: string
      ttl:
        type: string
      version:
        type: string
    type: object
...
  github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_search_response.Message:
    properties:
      catalog:
        $ref: '#/definitions/response.Catalog'
    type: object
...
  github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_select_request.Context:
    properties:
      action:
        type: string
      bap_id:
        type: string
      bap_uri:
        type: string
      bpp_id:
        type: string
      bpp_uri:
        type: string
      domain:
        type: string
      location:
        $ref: '#/definitions/github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_select_request.Location'
      message_id:
        type: string
      timestamp:
        type: string
      transaction_id:
        type: string
      ttl:
        type: string
      version:
        type: string
    type: object
...
  github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_select_request.Message:
    properties:
      order:
        $ref: '#/definitions/github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_select_request.Order'
    type: object
...
  github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_select_response.Message:
    properties:
      order:
        $ref: '#/definitions/github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_select_response.Order'
    type: object
...
  job.CreateJobRequest:
    properties:
      businessId:
        type: string
      description:
        type: string
      eligibility:
        $ref: '#/definitions/job.Eligibility'
      location:
        $ref: '#/definitions/job.Location'
      name:
        type: string
      salaryRange:
        $ref: '#/definitions/job.SalaryRange'
      type:
        $ref: '#/definitions/job.JobType'
      vacancies:
        type: integer
      workDays:
        $ref: '#/definitions/job.WorkDays'
      workHours:
        $ref: '#/definitions/job.WorkHours'
    type: object
...
  job.Eligibility:
    properties:
      academicQualification:
        $ref: '#/definitions/job.AcademicQualification'
      documentsRequired:
        items:
          $ref: '#/definitions/job.Document'
        type: array
      gender:
        $ref: '#/definitions/job.Gender'
      yearsOfExperience:
        type: integer
    type: object
  job.Gender:
    enum:
    - any
    - male
    - female
    type: string
    x-enum-varnames:
    - GenderAny
    - GenderMale
    - GenderFemale
  job.GetJobApplicationsResponse:
    properties:
      applicantDetails:
        $ref: '#/definitions/github_com_ONEST-Network_Job-Manager-Adapter_pkg_database_mongodb_job-application.ApplicantDetails'
      id:
        type: string
      status:
        $ref: '#/definitions/jobapplication.JobApplicationStatus'
    type: object
  job.JobType:
    enum:
    - full-time
    - part-time
    - contract
    - internship
    type: string
    x-enum-varnames:
    - JobTypeFullTime
    - JobTypePartTime
    - JobTypeContract
    - JobTypeInternship
  job.Location:
    properties:
      address:
        type: string
      areaCode:
        description: 'Postal code, for example: ''560102'''
        type: string
      city:
        description: 'STD code, for example: ''std:080'''
        type: string
      coordinates:
        $ref: '#/definitions/job.Coordinates'
      state:
        description: 'State code, for example: ''IN-KA'''
        type: string
      street:
        type: string
    type: object
  job.SalaryRange:
    properties:
      max:
        type: integer
      min:
        type: integer
    type: object
  job.WorkDays:
    properties:
      end:
        type: integer
      start:
        type: integer
    type: object
  job.WorkHours:
    properties:
      end:
        type: string
      start:
        type: string
    type: object
  jobapplication.JobApplicationStatus:
    enum:
    - APPLICATION_ACCEPTED
    - APPLICATION_REJECTED
    - ASSESSMENT_IN_PROGRESS
    - OFFER_REJECTED
    - OFFER_ACCEPTED
    - OFFER_EXTENDED
    - CANCELLED
    type: string
    x-enum-varnames:
    - JobApplicationStatusApplicationAccepted
    - JobApplicationStatusApplicationRejected
    - JobApplicationStatusAssessmentInProgress
    - JobApplicationStatusOfferRejected
    - JobApplicationStatusOfferAccepted
    - JobApplicationStatusOfferExtended
    - JobApplicationStatusCancelled
  jobapplication.UpdateJobApplicationStatusRequest:
    properties:
      status:
        description: '@Enum(APPLICATION_ACCEPTED, APPLICATION_REJECTED, ASSESSMENT_IN_PROGRESS,
          OFFER_REJECTED, OFFER_ACCEPTED, OFFER_EXTENDED, CANCELLED)'
        type: string
    type: object
...
  request.CancelRequest:
    properties:
      context:
        $ref: '#/definitions/github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_cancel_request.Context'
      message:
        $ref: '#/definitions/github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_cancel_request.Message'
    type: object
  request.ConfirmRequest:
    properties:
      context:
        $ref: '#/definitions/github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_confirm_request.Context'
      message:
        $ref: '#/definitions/github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_confirm_request.Message'
    type: object
...
  request.InitRequest:
    properties:
      context:
        $ref: '#/definitions/github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_init_request.Context'
      message:
        $ref: '#/definitions/github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_init_request.Message'
    type: object
...
  request.SearchRequest:
    properties:
      context:
        $ref: '#/definitions/github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_search_request.Context'
      message:
        $ref: '#/definitions/github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_search_request.Message'
    type: object
  request.SelectRequest:
    properties:
      context:
        $ref: '#/definitions/github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_select_request.Context'
      message:
        $ref: '#/definitions/github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_select_request.Message'
    type: object
...
  request.StatusRequest:
    properties:
      context:
        $ref: '#/definitions/github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_status_request.Context'
      message:
        $ref: '#/definitions/github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_status_request.Message'
    type: object
...
  response.CancelResponse:
    properties:
      context:
        $ref: '#/definitions/github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_cancel_response.Context'
      message:
        $ref: '#/definitions/github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_cancel_response.Message'
    type: object
...
  response.ConfirmResponse:
    properties:
      context:
        $ref: '#/definitions/github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_confirm_response.Context'
      message:
        $ref: '#/definitions/github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_confirm_response.Message'
    type: object
...
  response.InitResponse:
    properties:
      context:
        $ref: '#/definitions/github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_init_response.Context'
      message:
        $ref: '#/definitions/github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_init_response.Message'
    type: object
...
  response.SearchResponse:
    properties:
      context:
        $ref: '#/definitions/github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_search_response.Context'
      message:
        $ref: '#/definitions/github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_search_response.Message'
    type: object
  response.SelectResponse:
    properties:
      context:
        $ref: '#/definitions/github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_select_response.Context'
      message:
        $ref: '#/definitions/github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_select_response.Message'
    type: object
  response.StatusResponse:
    properties:
      context:
        $ref: '#/definitions/github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_status_response.Context'
      message:
        $ref: '#/definitions/github_com_ONEST-Network_Job-Manager-Adapter_pkg_types_payload_onest_status_response.Message'
    type: object
...
info:
  contact: {}
paths:
  /business/{id}/jobs:
    get:
      consumes:
      - application/json
      description: List jobs for a business
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/business.ListJobsResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: List Jobs
      tags:
      - Business
  /business/add:
    post:
      consumes:
      - application/json
      description: Add a business
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/business.AddBusinessRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Add business
      tags:
      - Business
  /cancel:
    post:
      consumes:
      - application/json
      description: Cancel job application
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.CancelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.CancelResponse'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Cancel job application
      tags:
      - ONEST Network
  /confirm:
    post:
      consumes:
      - application/json
      description: Confirm job application submission
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.ConfirmRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ConfirmResponse'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Confirm job application submission
      tags:
      - ONEST Network
  /init:
    post:
      consumes:
      - application/json
      description: Initialize job application
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.InitRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.InitResponse'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Initialize job application
      tags:
      - ONEST Network
  /job-application/{id}/status:
    post:
      consumes:
      - application/json
      description: Update a job application's status
      parameters:
      - description: Job Application ID
        in: path
        name: id
        required: true
        type: string
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/jobapplication.UpdateJobApplicationStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Update job application status
      tags:
      - Job Application
  /job/{id}/applications:
    get:
      consumes:
      - application/json
      description: Get job applications
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/job.GetJobApplicationsResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get job applications
      tags:
      - Job
  /job/create:
    post:
      consumes:
      - application/json
      description: Create a job posting
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/job.CreateJobRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create job
      tags:
      - Job
  /search:
    post:
      consumes:
      - application/json
      description: Send jobs
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.SearchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SearchResponse'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Send jobs
      tags:
      - ONEST Network
  /select:
    post:
      consumes:
      - application/json
      description: Send job fulfillment
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.SelectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SelectResponse'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Send job fulfillment
      tags:
      - ONEST Network
  /status:
    post:
      consumes:
      - application/json
      description: Send job application current status
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.StatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.StatusResponse'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Send job application current status
      tags:
      - ONEST Network
swagger: "2.0"